Create README.md
NM (Neural Module) or "Нейронный Модуль".
It is a scheme for the mindustry game.
It is a tool for simulation and training of neural networks (yes, right in the game).
NM needs a mod (infinite Processor V7) to work. It adds faster processors and roomy memory blocks.
NM has its own request-response language, thanks to which you can manage and configure NM.
NM uses the Forward-Forward algorithm (presented by Jeffrey Hinton at the NeurIPS 2022 conference)
to train neural networks, as its more popular counterpart is back propagation
it requires more memory and calculations. It would also be nice to experiment with this relatively new algorithm.
NM uses a conventional perceptron as the architecture of a neural network. Only 1 activation function is supported (Hyperbalic tangent).
This is my first project to reach such scales. Please don't be too strict.
In the future, it is planned to make NM2 With RNN support, Back propagation. Also in version 2, it is planned to add more memory and add "multithreading" by adding additional processors.